requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
python-http-client==3.3.1
pytz==2020.4
sendgrid==6.4.8
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
